+++ extensions/cornerstone/src/services/ViewportService/CornerstoneViewportService.ts
@@ -763,6 +763,12 @@ class CornerstoneViewportService extends PubSubService implements IViewportServi
       });
     }
 
+    // It's crucial not to return here because the volume may be loaded,
+    // but the viewport also needs to set the volume.
+    // if (!volumesNotLoaded.length) {
+    //   return;
+    // }
+
     // This returns the async continuation only
     return this.setVolumesForViewport(viewport, volumeInputArray, presentations);
   }
@@ -1114,8 +1120,8 @@ class CornerstoneViewportService extends PubSubService implements IViewportServi
 
     const { segmentationService } = this.servicesManager.services;
 
-    segmentationPresentation.forEach((presentation: SegmentationPresentationItem) => {
-      const { segmentationId, type, hydrated } = presentation;
+    segmentationPresentation.forEach((presentationItem: SegmentationPresentationItem) => {
+      const { segmentationId, type, hydrated } = presentationItem;
 
       if (hydrated) {
         segmentationService.addSegmentationRepresentation(viewport.id, {