+++ extensions/default/src/index.ts
@@ -35,6 +35,8 @@ import colorPickerDialog from './utils/colorPickerDialog';
 import promptSaveReport from './utils/promptSaveReport';
 import promptLabelAnnotation from './utils/promptLabelAnnotation';
 import usePatientInfo from './hooks/usePatientInfo';
+import { PanelStudyBrowserHeader } from './Panels/StudyBrowser/PanelStudyBrowserHeader';
+import * as utils from './utils';
 
 const defaultExtension: Types.Extensions.Extension = {
   /**
@@ -98,4 +100,6 @@ export {
   promptLabelAnnotation,
   colorPickerDialog,
   usePatientInfo,
+  PanelStudyBrowserHeader,
+  utils,
 };