+++ platform/ui-next/src/components/DisplaySetMessageListTooltip/DisplaySetMessageListTooltip.tsx
@@ -23,9 +23,9 @@ const DisplaySetMessageListTooltip = ({ messages, id }): React.ReactNode => {
             />
           </TooltipTrigger>
           <TooltipContent side="right">
-            <div className="max-w-68 text-left text-base text-white">
+            <div className="max-w-68 text-left text-lg text-white">
               <div
-                className="break-normal text-base font-semibold text-blue-300"
+                className="break-normal text-lg font-semibold text-blue-300"
                 style={{
                   marginLeft: '4px',
                   marginTop: '4px',